Refactor using each_with_object
You may have been tempted to use `inject` or `reduce`. The problem with
that approach is that you have to explicitly return the memo each time.
This isn't a problem when we're summing numbers, since we get a new
number after each reduction. However, it is a problem when we're
modifying an object.

See: [Better Hash Injection using each_with_object](http://technology.customink.com/blog/2014/10/14/better-hash-injection-using-each-with-object/)